Remove blocked label
We don’t need this label any more, as we track blocked stories using a column on our kanban board.

Because of the way this tool is currently set up, removing the label will not remove it from repos – we'll need to remove the label from each repo manually. However we still need to remove it otherwise it will be re-added the next time we change labels.