Add GitHub Actions
Adds workflows to test changes to this repo, and to sync any changes to
GitHub labels when changes are made to the main branch.

This replaces the functionality previously provided by Travis CI.

Instead of using the standard GitHub token when syncing labels, these
workflows use a personal access token generated by the
@govuk-design-system-ci machine user. The token is attached to this repo
as a secret with the name DESIGN_SYSTEM_GITHUB_TOKEN.

This commit also adds behaviour to post comments to PRs, showing the
output of `apply.sh` in dry-run mode, so developers/reviewers can easily
see the changes to labels that a PR will result in.